Split views into multiple files.
