+++ kagi/views/mixin.py
@@ -1,0 +1,5 @@
+class OriginMixin(object):
+    def get_origin(self):
+        return "{scheme}://{host}".format(
+            scheme=self.request.scheme, host=self.request.get_host()
+        )
